Move CLI code into a separate sub directory
Separate code that is related solely to the CLI application from the code that is related to the library itself (i.e. the engine).
This, together with the separation of the GUI code, is done not only for the purpose of separation, but to allow us to create a useful Python library for printing with Dymo printers.