Move library code into a separate sub directory
Separate code that is shared between CLI and GUI applications (i.e. the engine itself).
This, together with the separation of the GUI and UI code, is done not only for the purpose of separation, but to allow us to create a useful Python library for printing with Dymo printers.