Shorten exception text
When connecting a device, there's a short phase in which this error is shown.
Since it is very long, and is shown in the error widget, the whole app's layout is changed.
Instead, only show abbreviated message in the error label, and dump instructions to the console (we might want to have a dialog box for that in the GUI application in the future).