Major refactor for render engines
Have separate class for each kind of rendered (text, barcode, QR, etc.).
Combine renders are done using another kind of render (HorizontallyCombinedRenderEngine).
Move logic to associate renders and dymo device in a separate labeler_device file.
Note that the new architecture is design to support multiple labelers in the future.
This commit is the first step in that direction.